+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+##
+## Added
+- plugin description and useful links in admin settings
+
+## Changed
+- update Command service address
+- updated formats icons
+- display empty jwt warning if jwt secret is disabled
+
 ## 4.0.0
 ## Changed
 - compatible with HumHub 1.18